+++ samples/compose-showcase/src/commonMain/kotlin/com/awakekt/awake/sample/composeshowcase/app/ComposeShowcaseGame.kt
@@ -8,7 +8,6 @@ package com.awakekt.awake.sample.composeshowcase.app
 import com.awakekt.awake.engine.bootstrap.dsl.WindowDsl
 import com.awakekt.awake.engine.bootstrap.dsl.appDefinition
 import com.awakekt.awake.engine.bootstrap.dsl.appModule
-import com.awakekt.awake.engine.bootstrap.dsl.select
 import com.awakekt.awake.sample.composeshowcase.ui.ShowcaseApp
 import com.awakekt.awake.scene.authoring.scene
 
@@ -24,5 +23,4 @@ fun composeShowcase() = composeShowcaseDefinition.createApp()
 private fun WindowDsl.configureComposeShowcaseWindow() {
     title = "Awake Compose Showcase"
     size(1400, 900)
-    backend.select(platformBackendPreference())
 }