+++ samples/engine-showcase/src/commonMain/kotlin/com/awakekt/awake/showcase/app/EngineShowcaseApp.kt
@@ -7,7 +7,6 @@ package com.awakekt.awake.showcase.app
 
 import com.awakekt.awake.engine.bootstrap.dsl.WindowDsl
 import com.awakekt.awake.engine.bootstrap.dsl.appDefinition
-import com.awakekt.awake.engine.bootstrap.dsl.select
 import com.awakekt.awake.engine.platform.lifecycle.AwakeAppLifecycle
 import com.awakekt.awake.showcase.DEFAULT_SHOWCASE_ID
 import com.awakekt.awake.showcase.EngineShowcases
@@ -23,5 +22,4 @@ fun engineShowcaseApp(initialShowcaseId: String = DEFAULT_SHOWCASE_ID): AwakeApp
 private fun WindowDsl.configureShowcaseWindow() {
     title = "Awake Engine Showcase"
     size(1600, 900)
-    backend.select(platformBackendPreference())
 }