+++ samples/ui-showcase/src/commonMain/kotlin/com/awakekt/awake/sample/uishowcase/app/UiShowcaseGame.kt
@@ -7,7 +7,6 @@ package com.awakekt.awake.sample.uishowcase.app
 
 import com.awakekt.awake.engine.bootstrap.dsl.WindowDsl
 import com.awakekt.awake.engine.bootstrap.dsl.appDefinition
-import com.awakekt.awake.engine.bootstrap.dsl.select
 import com.awakekt.awake.sample.uishowcase.state.UiShowcaseRuntimeState
 
 private val uiShowcaseDefinition = appDefinition(createState = ::UiShowcaseRuntimeState) {
@@ -22,5 +21,4 @@ fun uiShowcase() = uiShowcaseDefinition.createApp()
 private fun WindowDsl.configureUiShowcaseWindow() {
     title = "Awake UI Showcase"
     size(1600, 900)
-    backend.select(platformBackendPreference())
 }